+++ pandas/tests/frame/test_axis_select_reindex.py
@@ -488,24 +488,3 @@ class TestDataFrameSelectReindex:
         result = df2.reindex(midx)
         expected = pd.DataFrame({"a": [0, 1, 2, 3, 4, 5, 6, np.nan, 8]}, index=midx)
         tm.assert_frame_equal(result, expected)
-
-    @pytest.mark.parametrize(
-        "operation", ["__iadd__", "__isub__", "__imul__", "__ipow__"]
-    )
-    @pytest.mark.parametrize("inplace", [False, True])
-    def test_inplace_drop_and_operation(self, operation, inplace):
-        # GH 30484
-        df = pd.DataFrame({"x": range(5)})
-        expected = df.copy()
-        df["y"] = range(5)
-        y = df["y"]
-
-        with tm.assert_produces_warning(None):
-            if inplace:
-                df.drop("y", axis=1, inplace=inplace)
-            else:
-                df = df.drop("y", axis=1, inplace=inplace)
-
-            # Perform operation and check result
-            getattr(y, operation)(1)
-            tm.assert_frame_equal(df, expected)
