+++ pandas/tests/frame/test_constructors.py
@@ -2677,6 +2677,13 @@ class TestDataFrameConstructors:
         with pytest.raises(ValueError, match="Shape of passed values"):
             DataFrame(dti, index=range(4))
 
+    def test_frame_ctor_datetime64_column(self):
+        rng = date_range("1/1/2000 00:00:00", "1/1/2000 1:59:50", freq="10s")
+        dates = np.asarray(rng)
+
+        df = DataFrame({"A": np.random.randn(len(rng)), "B": dates})
+        assert np.issubdtype(df["B"].dtype, np.dtype("M8[ns]"))
+
 
 class TestDataFrameConstructorWithDatetimeTZ:
     def test_from_dict(self):
