+++ pandas/tests/frame/test_timeseries.py
@@ -1,18 +1,11 @@
 import numpy as np
 
 import pandas as pd
-from pandas import DataFrame, date_range, to_datetime
+from pandas import DataFrame, to_datetime
 import pandas._testing as tm
 
 
 class TestDataFrameTimeSeriesMethods:
-    def test_frame_ctor_datetime64_column(self):
-        rng = date_range("1/1/2000 00:00:00", "1/1/2000 1:59:50", freq="10s")
-        dates = np.asarray(rng)
-
-        df = DataFrame({"A": np.random.randn(len(rng)), "B": dates})
-        assert np.issubdtype(df["B"].dtype, np.dtype("M8[ns]"))
-
     def test_frame_append_datetime64_col_other_units(self):
         n = 100
